minor fixes

+++ src/Playlist.sol
@@ -7,7 +7,7 @@ import "./libraries/TransferHelper.sol";
 contract Playlist is ERC1155 {
     struct Royalty {
         uint24 id;
-        uint256 amount;
+        uint64 amount;
     }
 
     /// Maximum number of playlists uint24 = 16,777,215;
@@ -17,20 +17,20 @@ contract Playlist is ERC1155 {
     /// Payment token
     address public currency;
     /// Monthly plan
-    uint256 public plan = 4 * 1e18;
+    uint64 public plan = 4 * 1e18;
     /// Maximum royalties paid per month
-    uint256 private maxAmount = 3 * 1e18;
+    uint64 private maxAmount = 3 * 1e18;
     /// OpenBeats fee
-    uint256 public fee = 1 * 1e18;
+    uint64 public fee = 1 * 1e18;
     /// Total feesEarned
-    uint256 private feesEarned;
-    uint256 public royaltyLength = 30;
+    uint96 private feesEarned;
+    uint8 public royaltyLength = 30;
 
     constructor(address _currency) ERC1155("https://api.openbeats.xyz/openbeats/v1/playlist/getbyid/{id}") {
         currency = _currency;
     }
 
-    function mint(uint256 id, uint256 supply) public {
+    function mint(uint24 id, uint24 supply) public {
         super._mint(_msgSender(), id, supply, "");
     }
 
@@ -39,10 +39,10 @@ contract Playlist is ERC1155 {
     }
 
     function payPlan(address from, Royalty[] calldata royalties) public {
-        uint256 _maxAmount;
+        uint64 _maxAmount;
         require(royalties.length <= royaltyLength, "Length");
 
-        for (uint256 i = 0; i < royalties.length; i++) {
+        for (uint8 i = 0; i < royalties.length; i++) {
             unchecked {
                 _maxAmount += royalties[i].amount;
             }
@@ -53,7 +53,7 @@ contract Playlist is ERC1155 {
         unchecked {
             feesEarned += fee;
         }
-        for (uint256 i = 0; i < royalties.length; i++) {
+        for (uint8 i = 0; i < royalties.length; i++) {
             /// Cannot overflow because the sum of all playlist balances can't exceed the max uint256 value.
             unchecked {
                 balanceOfPlaylist[royalties[i].id] += royalties[i].amount;
